JBoss production deployment script
I wrote a script today which can be used for production deployments
from the command line.  Be sure to make use of the "at" command
to schedule one time cron jobs for after hours deployments.

This script is meant to complement test deployment automation.
Avoiding a web interface deployment (like test) is best for a
production environment.