Added a few checks and compatibility.
I added a few preflight tests which I felt were necessary for the
script to run.

Additionally, I added a few variables so that this script can be
directly compatible with both JBoss and Tomcat.  I feel like this
compatibility will allow the deploy script to play nice with other
app servers as well.