Bugfix: console ownership and multiple instances
We don't want /dev/null to have ownership taken to a particular
user if it is selected as the console.

Grepping for $TOMCAT_HOME is not enough in the case of two instances
having a similar name containing another instance.

e.g. /opt/tomcat and /opt/tomcat-node1.  grepping for /opt/tomcat
would show both instances.