stderr permission denied bugfix
Fixed bug with stderr output.  I found that when integrating the
deploy.sh script with remote deployments and the sudo command that
/dev/stderr has ownership associated with the originaly tty.  That
means that attempting to redirect command > /dev/stderr results in
permission denied and there's no meaningful output to indicate
this is the problem with exception for su - user and running the
command manually like echo "test" > /dev/stderr.  To work around
this problem rather than writing to the special device stderr
it is best to just redirect file descriptor 1 to file descriptor
2 for error output.